refactor: update notification package structure and imports
