+++ src/main/java/app/bottlenote/user/repository/JpaNotificationRepository.java
@@ -1,7 +1,7 @@
 package app.bottlenote.user.repository;
 
-import app.bottlenote.user.domain.Notification;
-import app.bottlenote.user.domain.NotificationRepository;
+import app.bottlenote.user_notification.domain.Notification;
+import app.bottlenote.user_notification.domain.NotificationRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
