+++ src/main/java/app/bottlenote/user_notification/domain/UserPushConfig.java
@@ -1,4 +1,4 @@
-package app.bottlenote.user.domain;
+package app.bottlenote.user_notification.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
